// src/app/models/equipo.model.ts
export interface Equipo {
  id?: number;
  nombreEquipo: string; 
  descripcion?: string;
  tipo?: string;
  modelo?: string;
  marca?: string;
  serie?: string;
  ubicacionDelEquipo?: string;
  utilizacion?: string;
  recibidoPor?: string;
  proveedor?: string;
  ordenDeCompra?: string;
  factura?: string;
  fechaDeCompra?: Date;
  fechaFinGarantia?: Date;
  garantia?: string;
  precio?: number;
  procesador?: string;
  memoriaRamGB?: number;
  almacenamientoGB?: number;
  tipoAlmacenamiento?: string;
  placaBase?: string;
  fuentePoderWatts?: number;
  tarjetaGrafica?: string;
  tieneTarjetaRed?: boolean;
  tieneTarjetaSonido?: boolean;
  gabinete?: string;
  perifericosEntrada?: string;
  perifericosSalida?: string;
  componentes?: string;
  accesorios?: string;
  sistemaOperativo?: string;
  versionSO?: string;
  driversInstalados?: string;
  programasInstalados?: string;
  utilidadesSistema?: string;
  direccionIP?: string;
  direccionMAC?: string;
  }